+++ temporian/__init__.py
@@ -85,7 +85,7 @@ from temporian.implementation.numpy.data.plotter import plot
 from temporian.core.compilation import compile
 
 # Types
-from temporian.core.typing import EventSetOrNode
+from temporian.core import types
 
 # Runtime check
 from temporian.utils.typecheck import runtime_check_raise_exception