+++ temporian/implementation/numpy/operators/__init__.py
@@ -59,12 +59,13 @@ from temporian.implementation.numpy.operators import begin
 from temporian.implementation.numpy.operators import end
 from temporian.implementation.numpy.operators import enumerate
 from temporian.implementation.numpy.operators import fast_fourier_transform
+from temporian.implementation.numpy.operators import filter_moving_count
+from temporian.implementation.numpy.operators import map
 from temporian.implementation.numpy.operators import select_index_values
 from temporian.implementation.numpy.operators import since_last
 from temporian.implementation.numpy.operators import tick
 from temporian.implementation.numpy.operators import tick_calendar
 from temporian.implementation.numpy.operators import timestamps
 from temporian.implementation.numpy.operators import unique_timestamps
-from temporian.implementation.numpy.operators import filter_moving_count
 from temporian.implementation.numpy.operators import until_next
 from temporian.implementation.numpy.operators import where