+++ temporian/implementation/numpy/test/registered_operators_test.py
@@ -67,6 +67,7 @@ class RegisteredOperatorsTest(absltest.TestCase):
             "LESS_EQUAL_SCALAR",
             "LESS_SCALAR",
             "LOG",
+            "MAP",
             "MODULO",
             "MODULO_SCALAR",
             "MOVING_COUNT",